Refactor transform group resolvers: remove ICarriedTransformGroupResolver and...
