+++ src/Core.cs
@@ -1,6 +1,4 @@
-using CarryOn.API.Common;
 using CarryOn.API.Common.Interfaces;
-using CarryOn.Utility;
 using Vintagestory.API.Common;
 
 namespace CarryOn.CarryOnLib
@@ -11,7 +9,6 @@ namespace CarryOn.CarryOnLib
 
         public override void Dispose()
         {
-            Extensions.ClearCachedCarryManager();
             base.Dispose();
         }
     }
