+++ src/Utility/JsonHelper.cs
@@ -347,112 +347,5 @@ namespace CarryOn.Utility
             return map;
         }
 
-        public static ModelTransform GetTransform(JsonObject json, ModelTransform? baseTransform)
-        {
-            var trans = baseTransform?.Clone() ?? new ModelTransform();
-            if (TryGetVec3f(json, "translation", out var t))
-            {
-                trans.Translation = t;
-            }
-            else if (baseTransform != null)
-            {
-                trans.Translation = baseTransform.Translation;
-            }
-            if (TryGetVec3f(json, "rotation", out var r))
-            {
-                trans.Rotation = r;
-            }
-            else if (baseTransform != null)
-            {
-                trans.Rotation = baseTransform.Rotation;
-            }
-            if (TryGetVec3f(json, "origin", out var o))
-            {
-                trans.Origin = o;
-            }
-            else if (baseTransform != null)
-            {
-                trans.Origin = baseTransform.Origin;
-            }
-            // Try to get scale both as a Vec3f and single float - for compatibility reasons.
-            var hasVecScale = TryGetVec3f(json, "scale", out var sv);
-            var hasFloatScale = TryGetFloat(json, "scale", out var sf);
-            if (hasVecScale)
-            {
-                trans.ScaleXYZ = sv;
-            }
-            if (hasFloatScale)
-            {
-                trans.ScaleXYZ = new Vec3f(sf, sf, sf);
-            }
-            else if (!hasVecScale && baseTransform != null)
-            {
-                trans.ScaleXYZ = baseTransform.ScaleXYZ;
-            }
-
-            // Apply per-axis overrides on top of any full vector values.
-            if (TryGetFloat(json, "translationX", out var tx))
-            {
-                trans.Translation.X = tx;
-            }
-            if (TryGetFloat(json, "translationY", out var ty))
-            {
-                trans.Translation.Y = ty;
-            }
-            if (TryGetFloat(json, "translationZ", out var tz))
-            {
-                trans.Translation.Z = tz;
-            }
-
-            if (TryGetFloat(json, "rotationX", out var rx))
-            {
-                trans.Rotation.X = rx;
-            }
-            if (TryGetFloat(json, "rotationY", out var ry))
-            {
-                trans.Rotation.Y = ry;
-            }
-            if (TryGetFloat(json, "rotationZ", out var rz))
-            {
-                trans.Rotation.Z = rz;
-            }
-
-            if (TryGetFloat(json, "scaleX", out var sx))
-            {
-                trans.ScaleXYZ.X = sx;
-            }
-            if (TryGetFloat(json, "scaleY", out var sy))
-            {
-                trans.ScaleXYZ.Y = sy;
-            }
-            if (TryGetFloat(json, "scaleZ", out var sz))
-            {
-                trans.ScaleXYZ.Z = sz;
-            }
-
-            return trans;
-        }
-        
-        public static bool HasAnyTransformValue(JsonObject json)
-        {
-            if (json == null || !json.Exists)
-            {
-                return false;
-            }
-
-            return json.KeyExists("translation")
-                || json.KeyExists("rotation")
-                || json.KeyExists("origin")
-                || json.KeyExists("scale")
-                || json.KeyExists("translationX")
-                || json.KeyExists("translationY")
-                || json.KeyExists("translationZ")
-                || json.KeyExists("rotationX")
-                || json.KeyExists("rotationY")
-                || json.KeyExists("rotationZ")
-                || json.KeyExists("scaleX")
-                || json.KeyExists("scaleY")
-                || json.KeyExists("scaleZ");
-        }        
     }
 }
